pyproject.toml
[tool.poetry]
name = "reflexio"
version = "0.1.2"
description = "A Python library for the Reflexio"
authors = ["Reflexio Team"]
readme = "README.md"
packages = [{include = "reflexio"}]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.1"
psycopg2-binary = "^2.9.9"
boto3 = "^1.34.149"
sqlalchemy = "^2.0.31"
python-jose = "^3.3.0"
passlib = "^1.7.4"
openai = ">=2.8.0"
tenacity = "^9.0.0"
pre-commit = "^4.0.1"
duckduckgo-search = "^7.0.1"
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
black = "^24.10.0"
bcrypt = "^4.2.1"
pygwalker = "^0.4.9.13"
extra-streamlit-components = "^0.1.71"
reflexio-commons = {path = "reflexio/reflexio_commons", develop = true}
uvicorn = "^0.34.0"
xlsxwriter = "^3.2.2"
moto = "^5.0.28"
aiohttp = "^3.11.12"
supabase = "^2.15.1"
python-dotenv = "^1.1.0"
hdbscan = "^0.8.40"
streamlit-javascript = "^0.1.5"
redis = "^6.2.0"
websocket-client = "^1.8.0"
jupyterlab = "^4.4.3"
tiktoken = "^0.12.0"
anthropic = "^0.72.0"
litellm = "^1.80.11"
slowapi = "^0.1.9"
cachetools = "^6.2.4"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.5.3"
mkdocstrings = {extras = ["python"], version = "^0.24.0"}
griffe = "^0.38.0"
mkdocstrings-python = "^1.7.0"


[tool.poetry.group.dev.dependencies]
pytest-xdist = "^3.8.0"

[tool.pytest.ini_options]
# Parallel test execution with pytest-xdist
# -n auto: Use all available CPU cores
# --dist loadfile: Group tests by file (reduces fixture overhead, safer for in-file dependencies)
addopts = "-n auto --dist loadfile"
testpaths = ["reflexio/tests"]
asyncio_default_fixture_loop_scope = "function"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
